+++ tests/core/test_auth_email.py
@@ -7,7 +7,8 @@ import time_machine
 from canaille.app import mask_email
 from canaille.core.auth import AuthenticationSession
 from canaille.core.mails import send_one_time_password_mail
-from canaille.core.models import OTP_VALIDITY
+
+# OTP_VALIDITY now configured via OTP_LIFETIME
 from canaille.core.models import SEND_NEW_OTP_DELAY
 
 
@@ -135,7 +136,7 @@ def test_expired_email_otp(testclient, user, caplog):
         send_one_time_password_mail(user.preferred_email, otp)
         res.form["otp"] = user.one_time_password
 
-        traveller.shift(datetime.timedelta(seconds=OTP_VALIDITY))
+        traveller.shift(testclient.app.config["CANAILLE"]["OTP_LIFETIME"])
         res = res.form.submit(status=200, name="action", value="confirm")
 
         assert (